Dropped 4chan due to legitimate usage interference
Moot is gone, and while the staff are still cuckolds for hot pockets and /v/ is lost, I keep having to turn this script off to use other parts of the site not infected by SJWs. /vr/, /wg/ and other boards aren't SJW boards. It's only /v/ and /s4s/. 4chans inclusion on this list is ridiculous.|

This extension as an archiving tool for 4chan is inappropriate as well, as Anon Biblotechia does better work at that.